[FIX] Registry-lookup PR A1: реальний FK на CellValue.ValueRegistryEntryId (Q-222 закрито, Q-316) (#234)
CellValue.ValueRegistryEntryId (int?) не мав спільного CLR-типу з
RegistryEntry.Id (long), тож EF не міг зіставити зовнішній ключ, попри те
що фізично обидва зберігаються як int. Розширено до long? з
HasConversion<int?>() (той самий прийом, що MethodologyConstant.SubstanceEntryId)
і додано реальний FK_CellValue_Entry (RESTRICT) новою міграцією.

Компілятор провів через ~19 файлів без додаткових правок; вручну знайдено
й виправлено два місця сирого ADO повз конвеєр EF на найгарячішому шляху
системи (BulkCellLoader, NormalizedCellStore), де без явного звуження
SqlBulkCopy/SqlParameter отримали б boxed long проти фізичної int-колонки.

Три тести ConsistencyCheckJob*, що навмисно вставляли осиротілу комірку
для перевірки нічного бекстопу, адаптовано (NOCHECK/WITH NOCHECK CHECK
CONSTRAINT у try/finally — прецедент уже стояв у тому самому файлі для
FK_TableRow_Instance), не видалено.

+++ src/Ecr.Domain/ValueObjects/CellValueData.cs
@@ -14,7 +14,7 @@ public sealed record CellValueData
     public decimal? ValueNumeric { get; init; }
     public DateTime? ValueDate { get; init; }
     public bool? ValueBool { get; init; }
-    public int? ValueRegistryEntryId { get; init; }
+    public long? ValueRegistryEntryId { get; init; }
 
     /// <summary>Одиниця вимірювання; лише для <see cref="CellDataType.Unit"/> (R-A4).</summary>
     public int? ValueUnitId { get; init; }
